Fix printing issues
Add a Printing Options pane to the configure view dialog for all
views that support printing.
Update print dialog if number of pages changes. (when print options are changed)
Add print support for Accounts editor.
Do not create a painter on the printer (do not call painter()) as the
save()/restore() in the KoPrintingDialog class is strange.
Set page layout stuff for the reports. (There are issues with koreport here)

TODO: QPrintPreviewDialog freezes if more than one page is fed to it.