Fix (version) typos and license naming in Lesser GPL headers
There are a few issues with many Lesser GPL headers currently:
* refer to version "2" while there is only version "2.1" as minimum
* refer to "GPL" for more info, not "Lesser GPL"
* typos & possibly unwanted sideeffects of replace-all actions

This patch fixes those things. Quite a lot of files changed, so not fun to review, but you will be the hero of the day if you do.  At least I had some fun with perl and regular expressions, at least when it worked :)

See also http://lists.kde.org/?l=kde-licensing&m=142039469120716&w=2

REVIEW: 121996

Thanks to boud and ingwa for review!