Integrate TextDocumentInspector into textshape plugin for non-release builds
Needs to be explicitely enabled via entry in settings file of an app (like
wordsrc) to be activated, e.g.
[Debug]
EnableTextDocumentInspector=true

Blacklisting via DockerPluginsDisabled no longer works.

TextDocumentInspector needs to be in the same binary, due to using private
API of the textshape.
The old approach of a separate "plugin" via a separate desktop file with
a keyword for the plugin to be used with the kpluginfactory no longer works,
due to embedded plugin metadata.