Ensure we have app icons on Windows
While this is suboptimal, it seems there is some kind of bug when
mixing ecm_add_app_icon and kf5_add_kdeinit_executable which means
we end up with not having an icon set on Windows. This should be
fixed elsewhere, and is very much a hack, but this works for now.