Improve progress dialog
The dialog was difficult to understand and also too restrictive when disabling certain actions, see
BUG: 342574

Main change is to put both "resouce entry" tab and "date  entry" tab on the same page, which makes it easier to see the connection between them.
The "resource entry" part  is only shown when in "Per resource" edit mode.
When effort is entered in "resource entry" part, a date entry is automatically added in the "date entry" part.
Also there are now only two modes, "Per Resource" and "Per Task".
"Per Task" includes the functionallity in the prevoius "Calculate Effort" mode.
The default is now set to "Per Resource" as this preferred mode if PlanWork is used, so that this work by default.
If the manager wants easier data entry, he can select "Per Task" edit mode.

Tooltips and What's this has been added to further enhance usability.

FIXED-IN: 3.1.89